Rename new-game-master
